Added tile-Slot match checking, and puzzle progress
- Added tile checking; now the tiles places in the puzzle will check whether that is their correct slot or not.
- Added Puzzle Progress; upon placing a tile in its correct slot, the progress UI will be updated
- Fixed an issue where selecting a tile from the inventory (ScrollView) will put the tile behind it, and hid ut behind its background.

+++ Patterns Puzzle/Assets/PatternsPuzzle/Scripts/Controllers/UIController.cs
@@ -6,10 +6,10 @@ namespace GameControllers {
         public static UIController Instance => instance ??= FindObjectOfType<UIController>();
         private static UIController instance;
 
-        public static Action<float> OnProgressChanged;
+        public static Action<int> OnProgressChanged;
 
 
-        public void UpdateProgress(float progress) {
+        public void UpdateProgress(int progress) {
             OnProgressChanged?.Invoke(progress);
         }
     }
